重构详情界面
