+++ lib/pages/cartpage/empty_cart.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter_shop/provides/page_provide.dart';
+import 'package:flutter_shop/provides/page_index_provide.dart';
 import 'package:provide/provide.dart';
 
 class EmptyShoppingCart extends StatelessWidget {
