+++ lib/pages/index_page.dart
@@ -5,7 +5,7 @@ import 'package:flutter_shop/pages/cartpage/cart_page.dart';
 import 'package:flutter_shop/pages/category_page.dart';
 import 'package:flutter_shop/pages/homepage/home_page.dart';
 import 'package:flutter_shop/pages/mempage/mem_page.dart';
-import 'package:flutter_shop/provides/page_provide.dart';
+import 'package:flutter_shop/provides/page_index_provide.dart';
 import 'package:provide/provide.dart';
 
 class IndexPage extends StatelessWidget {
@@ -21,19 +21,18 @@ class IndexPage extends StatelessWidget {
 
     final _tabPages = <Widget>[HomePage(), CategoryPage(), CartPage(), MemPage()];
 
-    return Provide<PageIndexProvide>(builder: (_, child, pageProvide) {
-      return Scaffold(
-        backgroundColor: Color.fromRGBO(244, 245, 245, 1.0),
-        body: IndexedStack(index: pageProvide.page, children: _tabPages),
-        bottomNavigationBar: BottomNavigationBar(
-          items: _bottomTabs,
-          type: BottomNavigationBarType.fixed,
-          currentIndex: pageProvide.page,
-          onTap: (value) {
-            Provide.value<PageIndexProvide>(context).changePage(value);
-          },
-        ),
-      );
-    });
+    return Provide<PageIndexProvide>(
+        builder: (_, child, pageProvide) => Scaffold(
+              backgroundColor: Color.fromRGBO(244, 245, 245, 1.0),
+              body: IndexedStack(index: pageProvide.page, children: _tabPages),
+              bottomNavigationBar: BottomNavigationBar(
+                items: _bottomTabs,
+                type: BottomNavigationBarType.fixed,
+                currentIndex: pageProvide.page,
+                onTap: (value) {
+                  Provide.value<PageIndexProvide>(context).changePage(value);
+                },
+              ),
+            ));
   }
 }
