+++ lib/router/handler.dart
@@ -1,5 +1,5 @@
 import 'package:fluro/fluro.dart';
-import 'package:flutter_shop/pages/details_page.dart';
+import 'package:flutter_shop/pages/deatilspage/details_page.dart';
 import 'package:flutter_shop/pages/index_page.dart';
 
 Handler rootHandler = new Handler(handlerFunc: (_, params) => IndexPage());
